+++ src/com/udojava/evalex/ExpressionException.java
@@ -1,0 +1,11 @@
+/*
+    Created By : iamsubhranil
+    Date : 4/3/17
+    Time : 10:53 PM
+    Package : com.udojava.evalex
+    Project : GraphPlotter
+*/
+package com.udojava.evalex;
+
+public class ExpressionException {
+}
